# server/apis/gateway_api/tqmi-Gateway_API-1.0.0-swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "Gateway API",
    "description" : "Api for sending user info to server and to read user data"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
    "/user/info" : {
      "get" : {
        "tags" : [ "User" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Get the user information",
        "description" : "asd",
        "operationId" : "getUserInfo",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "applicatino/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "User" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "update user details in db",
        "description" : "update user details",
        "operationId" : "updateUser",
        "requestBody" : {
          "description" : "the users new details",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/institutions" : {
      "get" : {
        "tags" : [ "Institution" ],
        "summary" : "get lsit of institutions",
        "description" : "asd",
        "operationId" : "getInstitutions",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Institution"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/institution/{id}/templates" : {
      "get" : {
        "tags" : [ "Institution" ],
        "summary" : "get lsit of templates",
        "description" : "asd",
        "operationId" : "getInstitutionTemplates",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "description" : "institutions id",
          "schema" : {
            "type" : "string",
            "format" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FileResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put" : {
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Institution" ],
        "summary" : "upload template",
        "description" : "asd",
        "operationId" : "uploadInstitutionTemplate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "description" : "institutions id",
          "schema" : {
            "type" : "string",
            "format" : "string"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FileResponse"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/files/myFiles" : {
      "get" : {
        "tags" : [ "File" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Returns a list of filenames,authors and download links",
        "description" : "asd",
        "operationId" : "getMyFiles",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/FileResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "File" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Uploads the file to the stoarge bucket",
        "description" : "asd",
        "operationId" : "uploadFile",
        "requestBody" : {
          "required" : true,
          "description" : "The file",
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "filename" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "invalid request"
          }
        }
      }
    },
    "/files/{fileid}" : {
      "get" : {
        "tags" : [ "File" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Get file link",
        "description" : "Get the download link for the file",
        "operationId" : "getDownloadLink",
        "parameters" : [ {
          "name" : "fileid",
          "in" : "path",
          "required" : true,
          "description" : "id string that was sent with the file",
          "schema" : {
            "type" : "string",
            "format" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "format" : "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    },
    "schemas" : {
      "Institution" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "string"
          },
          "name" : {
            "type" : "string",
            "format" : "string"
          },
          "location" : {
            "type" : "string",
            "format" : "string"
          }
        }
      },
      "User" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "format" : "string"
          },
          "email" : {
            "type" : "string",
            "format" : "string"
          },
          "address" : {
            "type" : "string",
            "format" : "string"
          },
          "phone" : {
            "type" : "string",
            "format" : "string"
          },
          "accessLevel" : {
            "type" : "string",
            "format" : "string"
          }
        }
      },
      "FileResponse" : {
        "type" : "object",
        "properties" : {
          "fileName" : {
            "type" : "string",
            "format" : "string"
          },
          "owner" : {
            "type" : "string",
            "format" : "string"
          },
          "fileid" : {
            "type" : "string",
            "format" : "string"
          }
        }
      }
    }
  }
}